Renamed files
Renamed redundant parts of file names
