+++ .gitignore
@@ -1,1 +1,1 @@
-SEContestChatCommands.credentials.php
+credentials.php
