+++ SEContestChatCommands.php
@@ -5,16 +5,16 @@ $key = "random string";
 $channel = "SE channelID";
 $bearer = "SE JWT token";
 */
-require_once("SEContestChatCommands.credentials.php");
+require_once("credentials.php");
 
 if (isset($_GET["key"], $_GET["action"]) && $_GET["key"] == $key)
 {
-  require_once("SEContestChatCommands.open.php");
-  require_once("SEContestChatCommands.active.php");
-  require_once("SEContestChatCommands.close.php");
-  require_once("SEContestChatCommands.latest.php");
-  require_once("SEContestChatCommands.pick.php");
-  require_once("SEContestChatCommands.refund.php");
+  require_once("open.php");
+  require_once("active.php");
+  require_once("close.php");
+  require_once("latest.php");
+  require_once("pick.php");
+  require_once("refund.php");
 
   switch(strtolower($_GET["action"]))
   {
@@ -58,7 +58,7 @@ if (isset($_GET["key"], $_GET["action"]) && $_GET["key"] == $key)
 
     case "close":
       CloseContest($channel, $bearer, GetActiveId($channel, $bearer));
-      break; 
+      break;
 
     case "draw":
       if (isset($_GET["winner"]) && trim($_GET["winner"]) != "")
@@ -79,7 +79,7 @@ if (isset($_GET["key"], $_GET["action"]) && $_GET["key"] == $key)
 
     case "refund":
       RefundContest($channel, $bearer, GetLatestId($channel, $bearer));
-      break; 
+      break;
 
     default:
       echo "Invalid action! Expected: start|close|draw|refund";
